Move GatewayProtocol to tshd/gateway.ts
